+++ ui/mods/instant_sandbox/personalities.js
@@ -1,0 +1,7 @@
+define(['coui://ui/main/game/new_game/js/ai.js'], function() {
+  var personalities = ai_types()
+  for (var name in personalities) {
+    personalities[name].name = name
+  }
+  return personalities
+})